boop

+++ src/modules/richtext/richbio.tsx
@@ -1,15 +1,25 @@
-/* eslint-disable prettier/prettier */
 import React from "react";
 import { onClickOrKeyUp } from "~src/utils";
 import { bioWithoutImage, extractBioImage, saveBio } from "../altpfp";
 import { wrapRich } from "./richtext";
 
-function BioImage({ raw, onRemove }: { raw: string, onRemove: null | (() => void) }) {
-    const remove = onRemove && (() => {
-        if (confirm('Are you sure you want to remove your profile picture?'))
-            onRemove()
+type BioImageProps = {
+  raw: string;
+  onRemove: null | (() => void);
+};
+
+const confirmRemoval = () =>
+  confirm("Are you sure you want to remove your profile picture?");
+
+function BioImage({ raw, onRemove }: BioImageProps) {
+  const remove =
+    onRemove &&
+    (() => {
+      if (confirmRemoval()) onRemove();
     });
-    return <div style={{
+  return (
+    <div
+      style={{
         padding: "0.3em",
         margin: "0.5em",
         backgroundColor: "var(--ritsu-hair-color)",
@@ -18,53 +28,54 @@ function BioImage({ raw, onRemove }: { raw: string, onRemove: null | (() => void
         userSelect: "none",
         borderRadius: "0.3em",
         fontSize: "0.8em"
-    }}
-        title={raw}>
-        Profile picture ({raw.length} chars)
-        {!!remove &&
-            [
-                " ",
-                <b {...onClickOrKeyUp(remove)}>×</b>
-            ]}
-    </div >
+      }}
+      title={raw}
+    >
+      Profile picture ({raw.length} chars)
+      {!!remove && [" ", <b {...onClickOrKeyUp(remove)}>×</b>]}
+    </div>
+  );
 }
 
 export function init() {
-    UserProfile.prototype.bio = function bio() {
-        const {
-            compact_bio: compactBio,
-            data: { user }
-        } = this.state;
-        const isCompact = compactBio && user.bio.length > 202;
-        const bioImageRaw = extractBioImage(user.bio);
-        const bioStripped = bioWithoutImage(user.bio);
-        const bioImage = bioImageRaw && <BioImage raw={bioImageRaw} onRemove={
-            user.id === App.user.id
-                ? () => saveBio(user, bioStripped)
-                : null
-        } />;
-        // some icons may be cut off
-        const bioCompact = `${bioStripped.slice(0, 202)}... `;
-        const bioRaw = isCompact ? bioCompact : bioStripped;
-        return (
-            <span>
-                {wrapRich(bioRaw, (rest) => rest)}
-                {isCompact && (
-                    <span
-                        role="button"
-                        tabIndex={0}
-                        onMouseDown={() => this.bio_expand()}
-                        style={{
-                            color: '#00abf3',
-                            fontWeight: 'bold',
-                            cursor: 'pointer'
-                        }}
-                    >
-                        Show More
-                    </span>
-                )}
-                {!isCompact && bioImage}
-            </span>
-        );
-    };
+  UserProfile.prototype.bio = function bio() {
+    const {
+      compact_bio: compactBio,
+      data: { user }
+    } = this.state;
+    const isCompact = compactBio && user.bio.length > 202;
+    const bioImageRaw = extractBioImage(user.bio);
+    const bioStripped = bioWithoutImage(user.bio);
+    const bioImage = bioImageRaw && (
+      <BioImage
+        raw={bioImageRaw}
+        onRemove={
+          user.id === App.user.id ? () => saveBio(user, bioStripped) : null
+        }
+      />
+    );
+    // some icons may be cut off
+    const bioCompact = `${bioStripped.slice(0, 202)}... `;
+    const bioRaw = isCompact ? bioCompact : bioStripped;
+    return (
+      <span>
+        {wrapRich(bioRaw, (rest) => rest)}
+        {isCompact && (
+          <span
+            role="button"
+            tabIndex={0}
+            onMouseDown={() => this.bio_expand()}
+            style={{
+              color: "#00abf3",
+              fontWeight: "bold",
+              cursor: "pointer"
+            }}
+          >
+            Show More
+          </span>
+        )}
+        {!isCompact && bioImage}
+      </span>
+    );
+  };
 }
